requirements.txt
langchain 
faiss-cpu 
pypdf 
sentence-transformers
fastapi
uvicorn
pydantic
langchain-community
python-multipart
langchain-huggingface
accelerate>=0.26.0
pydantic
fastapi
google-generativeai 
# torch